+++ ReleaseNotes.md
@@ -34,6 +34,9 @@
 - [Technical] - Able to set PowerOn after a powerOff ( 0 for Off, 1 for On, 255, for previous state), Works for devices able to handle it. (Legrand, Philips)
 - [Technical] - Friendly Provisionning
 - [Technical] - Management of Zigate Blue Led. No need to restart plugin. Action taken immediatly from settings
+- [Technical] - Implement pingDevice policy (parameters pingDevices, pingDevicesFeq) . Ping every 3600s and once at plugin startup. The ping is much lighter than using the polling
+- [Technical] - Implement a way to force device status after command initiated. ( forcePollingAfterAction )
+
 
 ##             - 4.5.0 ( Web User Interface ) https://github.com/sasu-drooz/Domoticz-Zigate/releases/tag/4.5.0
 - [Technical] - Web Server Class